VectorTile import/export cleanup
